Added EditText object "itemPrice" Changed intPrice to itemPrice (type int to float) Changed price type int to float in the PostInformation.java

+++ app/src/main/java/com/example/myapplication/PostInformation.java
@@ -18,5 +18,5 @@ public class PostInformation {
     public String name;
     public String description;
     public String contact;
-    public int price;
+    public float price;
 }
